% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2partial_sim.R
\name{r2partial_sim}
\alias{r2partial_sim}
\title{Partial R-Squared Simulation for Logistic Regression}
\usage{
r2partial_sim(
  Q = Q,
  sample_size = sample_size,
  nsis = nsis,
  mediator_size = mediator_size,
  p = p,
  a = a,
  r = r,
  b = b,
  alpha1 = alpha1,
  Rsmethod = Rsmethod,
  nm2 = nm2,
  trainrate = trainrate,
  FDR = FDR,
  select = select,
  gamma = gamma,
  iter = iter,
  g = g,
  l = l
)
}
\arguments{
\item{Q}{Number of simulations to perform.}

\item{sample_size}{Number of samples per simulation.}

\item{nsis}{Parameter for Stepwise Information Criterion in mediator selection.}

\item{mediator_size}{Number of mediator variables.}

\item{p}{Total number of predictors.}

\item{a}{Coefficients for mediator interaction.}

\item{r}{Coefficient for the primary predictor.}

\item{b}{Coefficients for the mediators.}

\item{alpha1}{Intercept for the logistic model.}

\item{Rsmethod}{Method for calculating R-squared (e.g., "McFadden").}

\item{nm2}{Number of noise mediators.}

\item{trainrate}{Training set proportion.}

\item{FDR}{False Discovery Rate for selection.}

\item{select}{Flag to perform selection.}

\item{gamma}{Gamma parameter for penalization.}

\item{iter}{Iteration flag for selection process.}

\item{g}{Coefficient for additional predictors.}

\item{l}{Coefficient modeling the effect of Z on M.}
}
\value{
A list of simulation results, including R-squared values and selection metrics.
}
\description{
Conducts simulations to assess the partial R-squared values from logistic regression analyses, focusing on the contribution of selected covariates to the model's explanatory power.
}
\examples{
r2_results <- r2partial_sim(Q = 10)
}
