% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_fun.R
\name{sim_fun}
\alias{sim_fun}
\title{Simulate and Analyze Logistic Regression Data}
\usage{
sim_fun(
  Q = 500,
  sample_size = 2000,
  mediator_size = 2,
  p = 1000,
  a = rep(0, 2),
  r = 1,
  b = rep(0, 2),
  alpha1 = 0,
  Rsmethod = "Tjur",
  cov = NULL,
  g = g,
  l = l
)
}
\arguments{
\item{Q}{Number of simulations to run.}

\item{sample_size}{Number of samples per simulation.}

\item{mediator_size}{Number of mediator variables.}

\item{p}{Total number of predictors.}

\item{a}{Coefficients for mediators.}

\item{r}{Coefficient for the primary predictor.}

\item{b}{Coefficients for the mediators.}

\item{alpha1}{Intercept.}

\item{Rsmethod}{Method for calculating R-squared (e.g., "Tjur").}

\item{cov}{Optional covariance matrix.}

\item{g}{Additional parameter for `sim_logistic_cov`.}

\item{l}{Additional parameter for `sim_logistic_cov`.}
}
\value{
A list containing simulation results, including R-squared values, coefficients, and probabilities.
}
\description{
This function simulates logistic regression data and analyzes it using various models
to calculate R-squared values and other statistics. It can handle data with or without
a predefined covariance structure and allows for customization of simulation parameters.
}
\examples{
result <- sim_fun()
}
