% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_logistic_cov.R
\name{sim_logistic_cov}
\alias{sim_logistic_cov}
\title{Simulate Logistic Data with Covariates}
\usage{
sim_logistic_cov(
  sample_size = sample_size,
  mediator_size = mediator_size,
  p = p,
  a = a,
  r = r,
  b = b,
  alpha1 = alpha1,
  g = g,
  l = l
)
}
\arguments{
\item{sample_size}{Number of observations to generate.}

\item{mediator_size}{Number of mediators in the model.}

\item{p}{Total number of predictors.}

\item{a}{Coefficients affecting the relationship between X and M.}

\item{r}{Coefficient for the primary predictor X.}

\item{b}{Coefficients for mediators.}

\item{alpha1}{Intercept in the logistic regression equation.}

\item{g}{Coefficient for additional predictors.}

\item{l}{Coefficient modeling the effect of Z on M.}
}
\value{
A list containing the simulated dataset, mediator matrix, and probabilities.
}
\description{
Generates logistic regression data incorporating covariates' effects, allowing for complex relationships between predictors and the outcome variable.
}
\examples{
sim_data_cov <- sim_logistic_cov(sample_size = 100, mediator_size = 2, p = 5)
}
