% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_logistic_data.R
\name{sim_logistic_data}
\alias{sim_logistic_data}
\title{Simulate Logistic Data}
\usage{
sim_logistic_data(
  sample_size = 2000,
  mediator_size = 2,
  p = 1000,
  a = rep(0, 2),
  r = 0,
  b = rep(0, 2),
  alpha1 = 0,
  cov = NULL,
  g = g,
  l = l,
  nm1 = NULL
)
}
\arguments{
\item{sample_size}{Integer, the number of samples to generate.}

\item{mediator_size}{Integer, the number of mediator variables.}

\item{p}{Integer, the total number of predictors.}

\item{a}{Numeric vector, coefficients for mediators.}

\item{r}{Numeric, coefficient for the primary predictor.}

\item{b}{Numeric vector, coefficients for the mediators.}

\item{alpha1}{Numeric, intercept.}

\item{cov}{Optional covariance matrix.}

\item{g}{Additional parameters for `sim_logistic_cov`.}

\item{l}{Additional parameters for `sim_logistic_cov`.}

\item{nm1}{Optional, additional parameters for specifying extra noise mediators.}
}
\value{
A list containing the simulated dataset and additional simulation parameters.
}
\description{
This function simulates logistic regression data. It supports simulation with or without covariates and allows adjustment of various simulation parameters.
}
\examples{
sim_data <- sim_logistic_data(sample_size = 100, mediator_size = 2, p = 10)
}
